+++ CHANGELOG.rst
@@ -13,6 +13,10 @@ Bug fixes:
 
 - `make_object` is only called after all validators and postprocessors have finished (:issue:`253`). Thanks :user:`sunsongxp` for reporting.
 
+Other changes:
+
+- Change default error messages for ``URL`` and ``Email`` validators so that they don't include user input (:issue:`255`).
+
 2.0.0b4 (2015-07-07)
 ++++++++++++++++++++